moves Auth0Provider wrapper to index.js
