+++ src/components/NavBar.js
@@ -1,6 +1,8 @@
 import React from "react";
-import { useAuth0 } from "./auth/react-auth0-spa";
 import { Link } from "react-router-dom";
+
+// Auth0 imports
+import { useAuth0 } from "./auth/react-auth0-spa";
 
 const NavBar = () => {
   const { isAuthenticated, loginWithRedirect, logout } = useAuth0();
@@ -12,7 +14,7 @@ const NavBar = () => {
         <button aria-label="Access log in modal" onClick={() => loginWithRedirect({})}>Log in</button>
       )}
 
-      {/* If a user is logged in (authenticated), log out functionality will be enabled.) */}
+      {/* If a user is logged in (authenticated), log out button will show & functionality will be enabled.) */}
       {isAuthenticated && <button aria-label="log out" onClick={() => logout()}>Log out</button>}
 
       <Link to="/">Home</Link>
