+++ src/index.js
@@ -5,10 +5,30 @@ import App from './App';
 
 import * as serviceWorker from './serviceWorker';
 
+// Auth0 imports
+import { Auth0Provider } from "./components/auth/react-auth0-spa";
+import config from "./components/auth/auth_config.json";
+import history from "./utils/History";
 
+// Routes the user to dashboard upon login
+const onRedirectCallback = appState => {
+  history.push(
+    appState && appState.targetUrl
+      ? appState.targetUrl
+      : window.location.pathname
+  );
+};
 
-
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+  <Auth0Provider
+        domain={config.domain}
+        client_id={config.clientId}
+        redirect_uri={window.location.origin}
+        onRedirectCallback={onRedirectCallback}
+      >
+        <App />
+      </Auth0Provider>,
+      document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
